Re-organized tests.
