+++ tests/cargo-kani/stubbing-foreign-method/Cargo.toml
