+++ tests/cargo-kani/stubbing-foreign-method/expected
@@ -1,0 +1,4 @@
+error: Unable to find stub: other_crate::PubType::the_answer
+error: Unable to find stub: other_crate::PubType::the_answer
+error: Unable to find stub: other_crate::PrivType::the_answer
+error: could not compile `foreign-function-stubbing` due to 3 previous errors
