+++ tests/cargo-kani/stubbing-foreign-method/other_crate/Cargo.toml
