+++ tests/cargo-kani/stubbing-foreign-method/other_crate/src/lib.rs
@@ -1,22 +1,6 @@
 // Copyright Kani Contributors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-pub fn pub_fn() -> u32 {
-    0
-}
-
-pub fn fn_delegating_to_priv_fn() -> u32 {
-    priv_fn()
-}
-
-fn priv_fn() -> u32 {
-    0
-}
-
-fn the_answer() -> u32 {
-    42
-}
-
 pub struct PubType {}
 
 impl PubType {
@@ -58,25 +42,5 @@ impl PrivType {
 
     fn the_answer(&self) -> u32 {
         42
-    }
-}
-
-pub mod pub_mod {
-    pub fn pub_fn() -> u32 {
-        0
-    }
-
-    pub fn fn_delegating_to_priv_fn() -> u32 {
-        priv_mod::pub_fn()
-    }
-
-    mod priv_mod {
-        pub fn pub_fn() -> u32 {
-            0
-        }
-
-        fn the_answer() -> u32 {
-            42
-        }
     }
 }
