+++ tests/cargo-kani/stubbing-private-foreign-function/expected
@@ -1,0 +1,3 @@
+error: Unable to find stub: other_crate::pub_mod::priv_mod::the_answer
+error: Unable to find stub: other_crate::the_answer
+error: could not compile `foreign-function-stubbing` due to 2 previous errors
