+++ tests/cargo-kani/stubbing-public-foreign-function/expected
@@ -1,0 +1,1 @@
+VERIFICATION:- SUCCESSFUL
