+++ tests/cargo-kani/stubbing-public-foreign-function/other_crate/Cargo.toml
@@ -1,0 +1,8 @@
+# Copyright Kani Contributors
+# SPDX-License-Identifier: Apache-2.0 OR MIT
+[package]
+name = "other_crate"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
