+++ tests/cargo-kani/stubbing-public-foreign-function/other_crate/src/lib.rs
@@ -1,0 +1,12 @@
+// Copyright Kani Contributors
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+pub fn pub_fn() -> u32 {
+    0
+}
+
+pub mod pub_mod {
+    pub fn pub_fn() -> u32 {
+        0
+    }
+}
