+++ tests/ui/function-stubbing-error/expected
