+++ tests/ui/function-stubbing-error/main.rs
